Image processing/Image compression analyser/requirements.txt
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
imagehash>=4.3.1
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
argparse
Pillow-AVIF-Plugin>=1.0.0
webp>=0.1.0 